LayoutTests/http/tests/inspector/filesystem/read-file.html

<!DOCTYPE html>
<html>
<head>
<meta charset="UTF-8">
<script src="../inspector-test.js"></script>
<script src="filesystem-test.js"></script>
<script>
document.addEventListener("DOMContentLoaded", runTest);
function test()
{
    var requestManager = new WebInspector.FileSystemRequestManager();
    InspectorTest.clearFileSystem(step1);

    function step1()
    {
        InspectorTest.createDirectory("/hoge", step2);
    }

    function step2()
    {
        InspectorTest.writeFile("/hoge/fuga", "piPASSyo", step3);
    }

    function step3()
    {
        requestManager.requestFileContent("filesystem:http://127.0.0.1:8000/temporary/hoge/fuga", 2, 6, step4);
    }

    function step4(errorCode, content)
    {
        InspectorTest.addResult("errorCode = " + errorCode);
        if (content)
            InspectorTest.addResult("content = \"" + content + "\" [" + atob(content) + "]");
        else
            InspectorTest.addResult("content = (null)");

        InspectorTest.clearFileSystem(step5);
    }

    function step5()
    {
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body>
<p>Tests readFile.</p>
</body>
</html>
