LayoutTests/http/tests/inspector/filesystem/read-file-expected.txt

Tests requestFileContent.

errorCode = 0
content = "PASS"
charset = UTF-8
